+++ documentation/changelog.md
@@ -1,3 +1,8 @@
+# [v1.5.0] - 22/01/2024
+
+- Created donation button component in Footer
+- Add vercel Speed ​​Insights for monitoring
+
 # [v1.4.0] - 13/12/2023
 
 - Created contact page using emailjs